requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
pg8000==1.31.2
pydantic==2.8.2
python-dotenv==1.0.1
alembic==1.13.2
pytest==8.3.2
requests==2.32.3
flask-restx==1.3.0 